gantt chart for output viz

+++ solver.h
@@ -11,7 +11,7 @@ public:
     ~Solver();
     float calculatePriority(float te, int g);
     void updateWaitingTime(int casualty_id);
-    void updateCasualtyState(int casualty_id, int te);
+    void updateCasualtyState(int casualty_id, float te);
     void greedyAssignment();
 
     void printPriorityList();
